requirements.txt
Flask==2.3.3
Pillow==10.0.1
numpy==1.24.3
cryptography==41.0.4
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
cffi==1.15.1
pycparser==2.21